abstract to shared package
