+++ beacon-chain/sync/regular_sync_test.go
@@ -849,28 +849,3 @@ func TestHandleStateReq_OK(t *testing.T) {
 
 	testutil.AssertLogsContain(t, hook, "Sending beacon state to peer")
 }
-
-func TestSafelyHandleMessage(t *testing.T) {
-	hook := logTest.NewGlobal()
-
-	safelyHandleMessage(func(_ p2p.Message) {
-		panic("bad!")
-	}, p2p.Message{
-		Data: &pb.BeaconBlock{},
-	})
-
-	testutil.AssertLogsContain(t, hook, "Panicked when handling p2p message!")
-}
-
-func TestSafelyHandleMessage_NoData(t *testing.T) {
-	hook := logTest.NewGlobal()
-
-	safelyHandleMessage(func(_ p2p.Message) {
-		panic("bad!")
-	}, p2p.Message{})
-
-	entry := hook.LastEntry()
-	if entry.Data["msg"] != "message contains no data" {
-		t.Errorf("Message logged was not what was expected: %s", entry.Data["msg"])
-	}
-}
